Fix some data type stuff

+++ cmd/info.go
@@ -62,7 +62,7 @@ func infoSeq(in <-chan fastx.RecordChunk) <-chan *InfoRecord {
 					}
 				}
 
-				gcRatio := float64(gcBases) / float64(s.Length()-(nonATGCNBases+nBases))
+				gcRatio := float64(gcBases / (s.Length() - (nonATGCNBases + nBases)))
 
 				var qualScores int = 0
 				var errorProbs float64 = 0
@@ -80,7 +80,7 @@ func infoSeq(in <-chan fastx.RecordChunk) <-chan *InfoRecord {
 						qualScores += score
 						errorProbs += seqmath.ErrorProbForQ(score)
 
-						meanBaseQual = float64(qualScores) / float64(s.Length())
+						meanBaseQual = float64(qualScores / s.Length())
 						meanErrorProb = float64(errorProbs) / float64(s.Length())
 					}
 				}
@@ -217,16 +217,16 @@ specified.`,
 			}
 		}
 
-		totalGcRatio := float64(totalGcCount) / float64(totalSeqLength)
+		totalGcRatio := float64(totalGcCount / totalSeqLength)
 		totalGcPercent := totalGcRatio * 100
 
-		totalGcRatioNoAmbig := float64(totalGcCount) / float64(totalSeqLength-(totalNonATGCNBases+totalNBases))
+		totalGcRatioNoAmbig := float64(totalGcCount / (totalSeqLength - (totalNonATGCNBases + totalNBases)))
 		totalGcPercentNoAmbig := totalGcRatioNoAmbig * 100
 
 		meanQualityPerSeq := float64(sumMeanQualityScores) / float64(totalSeqs)
 		meanErrorProbPerSeq := float64(sumMeanErrorProbs) / float64(totalSeqs)
 
-		meanQualityPerBase := float64(sumBaseQualityScores) / float64(totalSeqLength)
+		meanQualityPerBase := float64(sumBaseQualityScores / totalSeqLength)
 		meanErrorProbPerBase := float64(sumBaseErrorProbs) / float64(totalSeqLength)
 
 		// NXX Calc
